[stable] Release 2.2.4
Fix bug in limit order logic around small orders

We discovered a minor bug whereby limit orders could be submitted that are so
small that they can be filled without the other side of the trade paying
anything. While this is generally harmless, some error-checking logic
made it so that new orders would be rejected if the order at the top of
the book had this issue. This change resolves this issue and adds
thorough testing around it.